front breakdown and migrations
