+++ aiden-web/aiden_app/templates/chat.html
@@ -33,7 +33,7 @@
     <div class="total-price-counter" id="total-price-counter">Total Price Paid: $0</div>
     <div class="profile-images"></div>
 
-    <script src="{% static 'js/main.js' %}"></script>
+    <script src="{% static 'js/main.js' %}" type="module"></script>
 </body>
 
 </html>
